third_party/blink/web_tests/external/wpt/mediasession/artwork-url-encoding-euc-kr-expected.txt

This is a testharness.js-based test.
[FAIL] artwork-url-encoding-euc-kr
  assert_equals: artwork src encoded with utf-8 expected "?icon=%EC%95%84%EC%9D%B4%EC%BD%98" but got "?icon=%BE%C6%C0%CC%C4%DC"
Harness: the test ran to completion.
